Implementando Padrão REST - 02

+++ src/main/java/com/example/libraryapi/service/AutorService.java
@@ -1,7 +1,10 @@
 package com.example.libraryapi.service;
 
+import com.example.libraryapi.exception.OperacaoNaoPermitidaException;
 import com.example.libraryapi.model.Autor;
 import com.example.libraryapi.repository.AutorRepository;
+import com.example.libraryapi.repository.LivroRepository;
+import com.example.libraryapi.validator.AutorValidator;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -12,12 +15,17 @@ import java.util.UUID;
 public class AutorService {
 
     private final AutorRepository autorRepository;
+    private final LivroRepository livroRepository;
+    private final AutorValidator autorValidator;
 
-    public AutorService(AutorRepository autorRepository) {
+    public AutorService(AutorRepository autorRepository, LivroRepository livroRepository, AutorValidator autorValidator) {
         this.autorRepository = autorRepository;
+        this.autorValidator = autorValidator;
+        this.livroRepository = livroRepository;
     }
 
     public Autor salvar(Autor autor){
+        autorValidator.validar(autor);
         return autorRepository.save(autor);
     }
 
@@ -29,10 +37,14 @@ public class AutorService {
         if(autor.getId()==null){
             throw new IllegalArgumentException("Para atualizar é necessário ter o autor registrado na base.");
         }
+        autorValidator.validar(autor);
         autorRepository.save(autor);
     }
 
     public void deletar(Autor autor) {
+        if(possuiLivro(autor)){
+            throw new OperacaoNaoPermitidaException("Não é permitido exclusão de autor que possui livros agregados.");
+        }
         autorRepository.delete(autor);
     }
 
@@ -47,6 +59,10 @@ public class AutorService {
             return autorRepository.findByNacionalidade(nacionalidade);
         }
         return autorRepository.findAll();
+    }
+
+    private boolean possuiLivro(Autor autor){
+        return livroRepository.existsByAutor(autor);
     }
 
 }
